+++ CHANGELOG.md
@@ -7,10 +7,15 @@ Changes on the `master` branch, but not yet released, will be listed here.
 ### Features
 
 -   `setNeedsUpdate` and `shouldUpdate` on `LayoutSource` now accepts a `IItemUpdateManyOptions` parameter. Multiple calls to `setNeedsUpdate` merges options and when `update` is called, the merged options are used.
+-   Added `linkLayout` on `EvergridLayout` to allow linking multiple grids in the x, y, ot both axes. This allows both grids to share target scroll information.
 
 ### Bug Fixes
 
 -   [[#26](https://github.com/diatche/evergrid/pull/26)] Fixed a bug where `scrollTo` would not work if there were certain side effects associated with desceleration.
+
+### Breaking Changes
+
+-   Removed `locationOffsetTarget$` and `scaleTarget$` animated properties from `EvergridLayout`. Use `willChangeLocationOffsetBase` and `willChangeScale` callbacks respectively.
 
 ## 0.2.0